+++ libs/Scene/src/scene/Scene.cpp
@@ -281,8 +281,8 @@ namespace SceneSystem {
 		Alabaster::EventDispatcher dispatch(event);
 		dispatch.dispatch<Alabaster::WindowResizeEvent>([this](Alabaster::WindowResizeEvent& e) {
 			const auto vertical_fov = glm::degrees(scene_camera->get_vertical_fov());
-
-			scene_camera.reset(new Alabaster::EditorCamera(vertical_fov, e.width(), e.height(), 0.1f, 1000.0f, scene_camera.get()));
+			scene_camera.reset(new Alabaster::EditorCamera(
+				vertical_fov, static_cast<float>(e.width()), static_cast<float>(e.height()), 0.1f, 1000.0f, scene_camera.get()));
 			this->scene_renderer->set_camera(scene_camera);
 			return false;
 		});
